Renamed App\Domain\Product to App\Domain\Product\Product
